% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphospace_functions.R
\name{add_noise}
\alias{add_noise}
\title{add_noise general wrapper}
\usage{
add_noise(MD, ...)
}
\arguments{
\item{MD}{an object to apply add_noise functions.}

\item{...}{additional parameters to be passed to add_noise functions.}
}
\value{
The output of an add_noise function given the method used.
}
\description{
Provides a general wrapper for add_noise functions.
}
\examples{
add_noise(MD, c(0.5, 0.5))

}
\author{
Sergei Tarasov
}
