% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphospace_functions.R
\name{add_noise_MD.md_1tree}
\alias{add_noise_MD.md_1tree}
\title{Adding noise to MDS from one stochastic character map}
\usage{
add_noise_MD.md_1tree(MD, add.noise)
}
\arguments{
\item{MD}{tibble. The output of a MD_morpho function.}

\item{add.noise}{numeric. A vector of length 2 indicating the amount of noise to be added to the x and y coordinates.}
}
\value{
A list of tibbles as in the output of MD_morpho functions, with noise added.
}
\description{
Adds noise to the points in the 2D coordinates in the MDS plot. 
# The noise is calculated as var(V)*add.noise.
}
\examples{
add_noise_MD.md_1tree(MD, c(0.5, 0.5))

}
\author{
Sergei Tarasov
}
