% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphospace_functions.R
\name{add_noise_MD}
\alias{add_noise_MD}
\title{Adding noise to MDS from one stochastic character map}
\usage{
add_noise_MD(MD, add.noise)
}
\arguments{
\item{MD}{tibble. The output of a MD_morpho function.}

\item{add.noise}{numeric. A vector of length 2 indicating the amount of noise to be added to the x and y coordinates.}
}
\value{
A list of tibbles as in the output of MD_morpho functions, with noise added.
}
\description{
Adds noise to the points in the 2D coordinates in the MDS plot. 
# The noise is calculated as var(V)*add.noise.
}
\examples{
data("hym_stm_mds")
# Get a sample of 5 amalgamated stochastic maps.
tree_list <- hym_stm_mds
tree_list <- merge_tree_cat_list(tree_list)
\dontrun{
  
  # Multidimensional scaling for an arbitrary tree.
  MD <- suppressWarnings(MultiScale.simmap(tree_list[[1]], add.noise = NULL))
  
  # Add noise.
  add_noise_MD(MD, c(0.3, 0.3))
  
}

}
\author{
Sergei Tarasov
}
