% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhpp_functions.R
\name{estimate_edge_KDE_Markov_kernel_unnorm}
\alias{estimate_edge_KDE_Markov_kernel_unnorm}
\title{Estimate the unnormalized Markov KDE}
\usage{
estimate_edge_KDE_Markov_kernel_unnorm(tree.discr, Path.data, h = 10)
}
\arguments{
\item{tree.discr}{simmap or phylo object. A discretized tree using the 'discr_Simmap' function.}

\item{Path.data}{numeric. A list of path data obtained from the 'make_data_NHPP_KDE_Markov_kernel' function.}

\item{h}{numeric. A value for the bandwidth calculated using the 'estimate_band_W' function.}
}
\value{
A list with the estimated unnormalized KDEs ($Maps.mean) for each edge.
}
\description{
Estimated the unnormalized Markov KDE for each edge averaged across all possible root-tip paths.
}
\examples{
Edge.KDE <- estimate_edge_KDE_Markov_kernel_unnorm(tree.discr, Pseudo.path.data, h = 15)

}
\author{
Sergei Tarasov
}
