% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhpp_functions.R
\name{get_path_edges}
\alias{get_path_edges}
\title{Get edges IDs from root to a given node.}
\usage{
get_path_edges(tree.merge, node)
}
\arguments{
\item{tree.merge}{simmap object.}

\item{node}{integer.

Internal function. Not exported.}
}
\description{
Get edges IDs from root to a given node.
}
\author{
Sergei Tarasov
}
