% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhpp_functions.R
\name{get_state}
\alias{get_state}
\title{Get state name for contMap plotting.}
\usage{
get_state(vec, x)
}
\arguments{
\item{vec}{numeric. A vector of intervals defining bins.}

\item{x}{numeric. A target value.}
}
\description{
Internal function. Not exported.
}
\author{
Diego Porto
}
