% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhpp_functions.R
\name{get_states_path}
\alias{get_states_path}
\title{Get state information about a given path.}
\usage{
get_states_path(tree.merge, node)
}
\arguments{
\item{tree.merge}{simmap object.}

\item{node}{integer.

Internal function. Not exported.}
}
\description{
Get state information about a given path.
}
\author{
Sergei Tarasov
}
