% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontophylo_data.R
\docType{data}
\name{HAO}
\alias{HAO}
\title{Hymenoptera Anatomy Ontology (HAO)}
\format{
List containing various ontological relationships and terms.
}
\usage{
HAO
}
\description{
The anatomy ontology of Hymenoptera (Yoder et al. 2010). This same ontology was also used in Tarasov et al. (2022).
}
\examples{
data(HAO)
}
\references{
Tarasov, S., Mikó, I. & Yoder, M.J. (2022) ontoFAST: an r package for interactive and semi-automatic annotation of characters with biological ontologies. Methods in Ecology and Evolution, 13, 324–329.
(\doi{10.1111/2041-210X.13753})

Yoder MJ, Mikó I, Seltmann KC, Bertone MA, Deans AR. 2010. A Gross Anatomy Ontology for Hymenoptera. PLoS ONE 5 (12): e15991.
(\doi{10.1371/journal.pone.0015991})

\href{http://portal.hymao.org/projects/32/public/ontology/}{Hymenoptera Anatomy Ontology Portal}
}
\keyword{datasets}
