% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontophylo_data.R
\docType{data}
\name{hym_annot}
\alias{hym_annot}
\title{Hymenoptera character annotations}
\format{
A data table with three columns and 30 rows; "char_id" contains character ids (e.g., "CH1", "CH2", "CH3");
"onto_id" contains ontology term ids from the HAO ontology (e.g., "HAO:0000234", "HAO:0000101", "HAO:0000639");
"label" contains the respective ontology term labels (e.g., "cranium", "antenna", "mouthparts").
Rows indicate characters.
}
\usage{
hym_annot
}
\description{
The character annotations from the first example data set used in the showcase analyses in Porto et al. (2023).
The annotations comprise the character ids, ontology term ids, and term labels used for each character.
}
\examples{
data(hym_annot) 
}
\references{
Porto, D.S., Uyeda, J., Mikó, I. & Tarasov, S. (2023) Supporting Data: ontophylo: Reconstructing the evolutionary dynamics of phenomes using new ontology-informed phylogenetic methods.
(\href{https://dx.doi.org/10.5281/zenodo.10285424}{Read})
}
\keyword{datasets}
