% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontophylo_data.R
\docType{data}
\name{hym_hm}
\alias{hym_hm}
\title{Hamming distances object (Hymenoptera example)}
\format{
A data table with 11 columns; "t.start" contains the starting times of mapped states; 
"t.end" contains the ending times of mapped states; "States" contains the amalgamated states; 
"Edge.id" contains the edge ids onto the tree; "delta.t" contains the duration of each mapped state;
"Ham.dist" contains the Hamming distances from the root state;
"Ham.dist.n" contains the normalized Hamming distances from the root state;
"Pois.count" contains the number of discrete changes;
"Focal.Edge.id", "tree.id", and "tree.tip.id" contain internal ids used by the package functions.
}
\usage{
hym_hm
}
\description{
Hamming distances data object from the amalgamated characters of "head"
obtained from the first example data set used in the showcase analyses in Porto et al. (2023).
The Hamming distances data object was obtained using the function "path_hamming_over_trees_KDE".
This data object is used to run the examples of the ontophyo package.
}
\examples{
data(hym_hm)
}
\references{
Porto, D.S., Uyeda, J., Mikó, I. & Tarasov, S. (2023) Supporting Data: ontophylo: Reconstructing the evolutionary dynamics of phenomes using new ontology-informed phylogenetic methods.
(\doi{10.5281/zenodo.10285424})
}
\keyword{datasets}
