% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontophylo_data.R
\docType{data}
\name{hym_img}
\alias{hym_img}
\title{Hymenoptera vector image}
\format{
A data object of class "grImport".
}
\usage{
hym_img
}
\description{
Example of a vector image of a hymenopteran wasp.
The original vector image in PostScript format was converted to XML format using (\href{https://www.ghostscript.com/}{GhostScript})
and the function "PostScriptTrace" from `grImport` (Murrell, 2009).
Then the XML file was imported into R using the function "readPicture" also from `grImport`.
This data object is used to run the examples of the ontophyo package.
}
\examples{
data(hym_img)
}
\references{
Murrell, P. (2009). Importing vector graphics: The grimport package for r. Journal of Statistical Software, 30:1–37.
(\href{https://doi.org/10.18637/jss.v030.i04}{Read})
}
\keyword{datasets}
