% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontophylo_data.R
\docType{data}
\name{hym_kde}
\alias{hym_kde}
\title{pNHPP rates object (Hymenoptera example)}
\format{
List containing pNHPP rates estimated for all edges of the tree sample.
}
\usage{
hym_kde
}
\description{
The data object contains pNHPP rates estimated for the amalgamated characters of "head"
obtained from the first example data set used in the showcase analyses in Porto et al. (2023).
The data object contains the estimated rates for all edges of the tree sample;
$Maps.mean and $Maps.mean.norm contain the raw and normalized rates estimated using the function "estimate_edge_KDE"; 
$Maps.mean.loess and $Maps.mean.loess.norm contain the smoothed raw and normalized rates estimated using the function "loess_smoothing_KDE";
$lambda.mean, $loess.lambda.mean, and $loess.lambda.mean.deriv contain the posteriors estimated for the raw and normalized rates, and its derivative.
This data object is used to run the examples of the ontophyo package.
}
\examples{
data(hym_kde)
}
\references{
Porto, D.S., Uyeda, J., Mikó, I. & Tarasov, S. (2023) Supporting Data: ontophylo: Reconstructing the evolutionary dynamics of phenomes using new ontology-informed phylogenetic methods.
(\doi{10.5281/zenodo.10285424})
}
\keyword{datasets}
