% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontophylo_data.R
\docType{data}
\name{hym_matrix}
\alias{hym_matrix}
\title{Hymenoptera character matrix}
\format{
A data table with 20 rows and 30 columns; each row indicates a species and each column a character.
}
\usage{
hym_matrix
}
\description{
The character matrix from the first example data set used in the showcase analyses in Porto et al. (2023).
The dataset comprises 30 simulated binary characters and the original matrix was reduced to contain only the 
20 representative taxa used for the package examples.
}
\examples{
data(hym_matrix) 
}
\references{
Porto, D.S., Uyeda, J., Mikó, I. & Tarasov, S. (2023) Supporting Data: ontophylo: Reconstructing the evolutionary dynamics of phenomes using new ontology-informed phylogenetic methods.
(\href{https://dx.doi.org/10.5281/zenodo.10285424}{Read})
}
\keyword{datasets}
