% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontophylo_data.R
\docType{data}
\name{hym_nhpp}
\alias{hym_nhpp}
\title{Changing times object (Hymenoptera example)}
\format{
List containing changing times between states for all edges of the tree sample.
}
\usage{
hym_nhpp
}
\description{
Changing times data object from the amalgamated characters of "head"
obtained from the first example data set used in the showcase analyses in Porto et al. (2023).
The changing times data object was obtained using the function "make_data_NHPP_KDE_Markov_kernel".
This data object is used to run the examples of the ontophyo package.
}
\examples{
data(hym_nhpp)
}
\references{
Porto, D.S., Uyeda, J., Mikó, I. & Tarasov, S. (2023) Supporting Data: ontophylo: Reconstructing the evolutionary dynamics of phenomes using new ontology-informed phylogenetic methods.
(\href{https://dx.doi.org/10.5281/zenodo.10285424}{Read})
}
\keyword{datasets}
