% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontophylo_data.R
\docType{data}
\name{hym_stm}
\alias{hym_stm}
\title{Hymenoptera stochastic character maps}
\format{
List containing 30 objects of class "multiSimmap".
}
\usage{
hym_stm
}
\description{
List of stochastic character maps obtained from all characters of 
the first example data set used in the showcase analyses in Porto et al. (2023). 
Only 50 samples per character were included to reduce file size (originally 100 samples).
}
\examples{
data(hym_stm)
}
\references{
Porto, D.S., Uyeda, J., Mikó, I. & Tarasov, S. (2023) Supporting Data: ontophylo: Reconstructing the evolutionary dynamics of phenomes using new ontology-informed phylogenetic methods.
(\href{https://dx.doi.org/10.5281/zenodo.10285424}{Read})
}
\keyword{datasets}
