% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontophylo_data.R
\docType{data}
\name{hym_stm_amalg}
\alias{hym_stm_amalg}
\title{Hymenoptera amalgamated stochastic character maps}
\format{
List containing three objects of class "multiPhylo".
}
\usage{
hym_stm_amalg
}
\description{
List of amalgamated stochastic character maps obtained from  
the first example data set used in the showcase analyses in Porto et al. (2023). 
Character were amalgamated into groups of 10 characters representing 
three anatomical regions of Hymenoptera: "head", "mesosoma", and "metasoma".
Only 50 samples per character were included to reduce file size (originally 100 samples).
}
\examples{
data(hym_stm_amalg)
}
\references{
Porto, D.S., Uyeda, J., Mikó, I. & Tarasov, S. (2023) Supporting Data: ontophylo: Reconstructing the evolutionary dynamics of phenomes using new ontology-informed phylogenetic methods.
(\href{https://dx.doi.org/10.5281/zenodo.10285424}{Read})
}
\keyword{datasets}
