% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontophylo_data.R
\docType{data}
\name{hym_stm_mds}
\alias{hym_stm_mds}
\title{Hymenoptera amalgamated phenome}
\format{
An stochastic character map of class "simmap".
}
\usage{
hym_stm_mds
}
\description{
Example of a stochastic character map obtained from the amalgamation of 394 characters
modified from the matrix of Sharkey et al. (2011) and using the tree from Klopfstein et al. (2013).
The tree was dated using penalized likelihood as implemented in TreePL (Smith & O’Meara, 2012).
This data object is used to run the examples of the ontophyo morphospace application.
}
\examples{
data(hym_stm_mds)
}
\references{
Sharkey, M.J., et al. 2011. Phylogenetic relationships among superfamilies of Hymenoptera. Cladistics 28(1), 80-112.
(\href{https://doi.org/10.1111/j.1096-0031.2011.00366.x}{Read})

Klopfstein, S., Vilhelmsen, L., Heraty, J.M., Sharkey, M. & Ronquist, F. (2013) The hymenopteran tree of life: evidence from protein-coding genes and objectively aligned ribosomal data. PLoS One, 8, e69344.
(\href{https://doi.org/10.1371/journal.pone.0069344}{Read})
}
\keyword{datasets}
