% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontophylo_data.R
\docType{data}
\name{hym_tree}
\alias{hym_tree}
\title{Hymenoptera dated tree}
\format{
A phylogenetic tree of class "phylo".
}
\usage{
hym_tree
}
\description{
A phylogenetic tree modified from Klopfstein et al. (2013).
The tree was dated using penalized likelihood as implemented in TreePL (Smith & O’Meara, 2012) and
then pruned to contain 20 representative taxa used for the package examples.
}
\examples{
data(hym_tree)
}
\references{
Klopfstein, S., Vilhelmsen, L., Heraty, J.M., Sharkey, M. & Ronquist, F. (2013) The hymenopteran tree of life: evidence from protein-coding genes and objectively aligned ribosomal data. PLoS One, 8, e69344.
(\href{https://doi.org/10.1371/journal.pone.0069344}{Read})
}
\keyword{datasets}
