% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhpp_functions.R
\name{KDE_unnorm_trunc_Markov}
\alias{KDE_unnorm_trunc_Markov}
\title{KDE for unnormalized Markov kernel vectorized.}
\usage{
KDE_unnorm_trunc_Markov(x, h, dat)
}
\arguments{
\item{x}{numeric. A vector with global changing times.}

\item{h}{numeric. The bandwidth value.

Internal function. Not exported.}

\item{dat}{numeric. A vector with path data values.}
}
\description{
KDE for unnormalized Markov kernel vectorized.
}
\author{
Sergei Tarasov
}
