% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhpp_functions.R
\name{KDE_unnormalized_scalar_Markov_kernel}
\alias{KDE_unnormalized_scalar_Markov_kernel}
\title{KDE for unnormalized Markov kernel.}
\usage{
KDE_unnormalized_scalar_Markov_kernel(x, h, dat)
}
\description{
Internal function. Not exported.
}
\author{
Sergei Tarasov
}
