% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhpp_functions.R
\name{loess_smoothing_KDE}
\alias{loess_smoothing_KDE}
\title{Get loess smoothing for the unnormalized Markov KDE}
\usage{
loess_smoothing_KDE(tree.discr, Edge.KDE)
}
\arguments{
\item{tree.discr}{simmap or phylo object. A discretized tree using the 'discr_Simmap' function.}

\item{Edge.KDE}{list. A list with the estimated unnormalized KDEs ($Maps.mean) for each edge.}
}
\value{
A list with the loess smoothing calculated for each edge.
}
\description{
Calculates loess smoothing for the unnormalized Markov KDE obtained from the 'estimate_edge_KDE_Markov_kernel_unnorm' function.
}
\examples{
edge.KDE$Maps.mean.loess <- loess_smooting_KDE(tree.discr, edge.KDE)

}
\author{
Sergei Tarasov
}
