% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhpp_functions.R
\name{make_contMap_KDE}
\alias{make_contMap_KDE}
\title{Make contMap KDE object}
\usage{
make_contMap_KDE(tree.discr, Edge.KDE.stat)
}
\arguments{
\item{tree.discr}{phylo object. A discretized tree using the 'discr_Simmap' function.}

\item{Edge.KDE.stat}{list. A list with the distributions of the estimated parameter of KDEs for each edge.}
}
\value{
A 'contMap' object.
}
\description{
Produces a contMap object for plotting the NHPP.
}
\examples{
NHPP.Map <- make_contMap(tree.discr, Edge.KDE$Maps.mean.norm)

}
\author{
Sergei Tarasov
}
