% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhpp_functions.R
\name{make_data_NHPP_over_edge_MarkovKDE}
\alias{make_data_NHPP_over_edge_MarkovKDE}
\title{Get NHPP data for a given edge (Markov KDE)}
\usage{
make_data_NHPP_over_edge_MarkovKDE(Tb.trees, Focal.Edge)
}
\arguments{
\item{Tb.trees}{data.frame. A tibble obtained with the 'path_hamming_over_trees_KDE' function.}

\item{Focal.Edge}{integer. A value indicating the edge ID.}
}
\value{
A numeric vector with changing times between states for a given edge.
}
\description{
Gets data on changing times between states for a given edge of a given sample of trees for the Markov kernel density estimator (KDE).
}
\examples{
make_data_NHPP_over_edge_MarkovKDE(Tb.trees, Focal.Edge = 5)

}
\author{
Sergei Tarasov
}
