% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhpp_functions.R
\name{make_postPois_KDE}
\alias{make_postPois_KDE}
\title{Make posterior distribution of NHPP}
\usage{
make_postPois_KDE(Edge.KDE.stat, lambda.post, lambda.post.stat = "Mean")
}
\arguments{
\item{Edge.KDE.stat}{list. A list with the estimated normalized or loess smoothing KDEs for each edge.}

\item{lambda.post}{list. A list with the distribution statistics calculated with the 'posterior_lambda_KDE' function.}

\item{lambda.post.stat}{character. A value with the statistic to be used.}
}
\value{
A list with the distribution of the selected statistic for each edge.
}
\description{
Produces a posterior distribution from a given list of statistics calculated with the 'posterior_lambda_KDE' function.
}
\examples{
make_postPois_KDE(Edge.KDE.stat, lambda.post, lambda.post.stat = 'Mean')

}
\author{
Sergei Tarasov
}
