% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphospace_functions.R
\name{MD_morpho}
\alias{MD_morpho}
\title{MDS general wrapper}
\usage{
MD_morpho(MD)
}
\arguments{
\item{MD}{an object to apply MultiScale functions.}
}
\value{
The output of a MultiScale function given the method used.
}
\description{
Provides a general wrapper for MDS functions.
}
\examples{
MD_morpho(tree.merge)

}
\author{
Sergei Tarasov
}
