% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphospace_functions.R
\name{mds_plot}
\alias{mds_plot}
\title{Plot morphospace from MDS}
\usage{
mds_plot(MD, Tslice = max(MD$Points$time))
}
\arguments{
\item{MD}{list. A list with the morphospace information obtained using the 'MultiScale.simmap' function.}

\item{Tslice}{numeric. The value for the temporal slices to be plotted, from root to tip. For example, if Tslice = 25, then all points in the morphospaces from time 0 (root) to 25 will be plotted.}
}
\description{
Wrapper function for plotting morphospaces obtained using the MultiScale.simmap' function.
}
\examples{
data("hym_stm_mds")
# Get a sample of 5 amalgamated stochastic maps.
tree_list <- hym_stm_mds
tree_list <- merge_tree_cat_list(tree_list)
# Multidimensional scaling for an arbitrary tree.
\dontrun{
  
  MD <- suppressWarnings(MultiScale.simmap(tree_list[[1]], add.noise = c(0.3,0.3)))

  MD_plot <- mds_plot(MD, Tslice = 10)
  MD_plot
  MD_plot <- mds_plot(MD, Tslice = 50)
  MD_plot
  MD_plot <- mds_plot(MD, Tslice = 200)
  MD_plot
  MD_plot <- mds_plot(MD, Tslice = 280)
  MD_plot
  
}

}
\author{
Diego Porto
}
