% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramo_functions.R
\name{merge_branch_cat}
\alias{merge_branch_cat}
\title{Merge state bins over branch}
\usage{
merge_branch_cat(br)
}
\arguments{
\item{br}{numeric or character vector. The branches of the tree.}
}
\value{
A numeric or character vector with merged identical bins.
}
\description{
Merges identical state bins over the same branch in the discretized stochastic map.
}
\examples{
tree <- data(HYM)$stm[[1]][[1]]
stm_discr <- discr_Simmap(tree, res = 100)
br <- stm_discr$maps[[50]]
br
merge_branch_cat(br)
stm_discr$maps <- lapply(stm_discr$maps, merge_branch_cat)
plot(stm_discr, setNames(getPalette(length(states)), states))

}
\author{
Sergei Tarasov
}
