% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramo_functions.R
\name{merge_tree_cat}
\alias{merge_tree_cat}
\title{Merge state bins over a tree}
\usage{
merge_tree_cat(tree)
}
\arguments{
\item{tree}{simmap object.}
}
\value{
A tree with merged identical bins.
}
\description{
Merges identical state bins over a tree in the discretized stochastic map.
}
\examples{
tree <- data(HYM)$stm[[1]][[1]]
tree <- discr_Simmap(tree)
stm_merg <- merge_tree_cat(tree)
stm_merg

}
\author{
Sergei Tarasov
}
