% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramo_functions.R
\name{paramo.list}
\alias{paramo.list}
\title{Stack multiple discrete stochastic character map lists}
\usage{
paramo.list(cc, tree.list, ntrees = 1)
}
\arguments{
\item{cc}{character. Characters IDs to stack.}

\item{tree.list}{multiSimmap or multiPhylo object. Named list with stochastic character maps.}

\item{ntrees}{integer. Number of trees to stack.}
}
\value{
A list of stacked stochastic character maps.
}
\description{
Performs the final stacking of maps for a set of stochastic character maps stored in a list.
}
\examples{
tree_list <- data(HYM)$stm[[1]]
tree_list <- discr_Simmap(tree_list)
tree_list <- merge_tree_cat_list(tree_list)
paramo.list(cc, tree_list, 10)

}
\author{
Sergei Tarasov
}
