% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhpp_functions.R
\name{path_hamming_over_trees_KDE}
\alias{path_hamming_over_trees_KDE}
\title{Hamming distances for a list of trees}
\usage{
path_hamming_over_trees_KDE(tree.list)
}
\arguments{
\item{tree.list}{multiSimmap object.}
}
\value{
A tibble with information on state changes, time spent on each state, edge IDs, absolute and normalized hamming distances for all edges and all trees in a list.
}
\description{
Calculates hamming distances for all paths in each discretized tree of a list.
}
\examples{
Tb.trees <- path_hamming_over_trees_KDE(tree.list)

}
\author{
Sergei Tarasov
}
