% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramo_functions.R
\name{read_Simmap_Rev}
\alias{read_Simmap_Rev}
\title{Reading stochastic character maps file from ReVBayes}
\usage{
read_Simmap_Rev(file, start = 1, end = 1, save = NULL)
}
\arguments{
\item{file}{character. Path to the RevBayes file.}

\item{start}{integer. First tree of the sample to start reading the RevBayes file.}

\item{end}{integer. Last tree of the sample to finish reading the RevBayes file.}

\item{save}{character. Name to save output file.}
}
\value{
A tree in 'phylip' format.
}
\description{
Imports stochastic character maps file from RevBayes into R.
}
\examples{
\dontrun{

  stm <- read_Simmap_Rev("revbayes.stm", start = 200, end = 500, save = NULL)
  stm <- phytools::read.simmap(text = stm, format = "phylip")
  phytools::plotSimmap(stm[[1]])

}

}
\author{
Sergei Tarasov
}
