% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramo_functions.R
\name{stack2}
\alias{stack2}
\title{Stack two discrete stochastic character map lists; x and y are the list of state names (i.e. maps).}
\usage{
stack2(x, y)
}
\arguments{
\item{x}{list. A list of state names.}

\item{y}{list. A list of state names.

Internal function. Not exported.}
}
\description{
Stack two discrete stochastic character map lists; x and y are the list of state names (i.e. maps).
}
\author{
Sergei Tarasov
}
