% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramo_functions.R
\name{stack_stm}
\alias{stack_stm}
\title{Stack two discrete stochastic character maps.}
\usage{
stack_stm(stm.list)
}
\arguments{
\item{stm.list}{list. A list of stochastic maps to be amalgamated.

Internal function. Not exported.}
}
\description{
Stack two discrete stochastic character maps.
}
\author{
Sergei Tarasov
}
