% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhpp_functions.R
\name{derivative_KDE}
\alias{derivative_KDE}
\title{Calculate KDE derivative over edges}
\usage{
derivative_KDE(tree.discr, Edge.KDE.stat)
}
\arguments{
\item{tree.discr}{simmap or phylo object. A discretized tree using the 'discr_Simmap' function.}

\item{Edge.KDE.stat}{list. A list with the estimated normalized or loess smoothing KDEs for each edge.}
}
\value{
A list with the distribution of the derivatives calculated for each edge.
}
\description{
Calculates the derivative of the normalized Markov KDE or normalized loess smoothing over edges.
}
\examples{
derivative_KDE(tree.discr, Edge.KDE.stat)

}
\author{
Sergei Tarasov
}
