% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramo_functions.R
\name{get_rough_state_cols}
\alias{get_rough_state_cols}
\title{Multiple character state colors}
\usage{
get_rough_state_cols(tree)
}
\arguments{
\item{tree}{simmap object.}
}
\value{
A character vector with colors associated with state names.
}
\description{
Get state colors for ploting stochastic character maps when there many states.
}
\examples{
library("RColorBrewer")
tree <- data(HYM)$stm[[1]][[1]]
plotSimmap(tree, get_rough_state_cols(tree),  lwd = 3, pts = F,ftype = "off", ylim = c(0,100))

}
\author{
Sergei Tarasov
}
