% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhpp_functions.R
\name{join_edges}
\alias{join_edges}
\title{Join neighboring edges in edge profiles.}
\usage{
join_edges(tree.discr, edge.profs)
}
\arguments{
\item{tree.discr}{phylo object. A discretized tree using the 'discr_Simmap' function.}

\item{edge.profs}{tibble. A tibble with edge profile information.}
}
\description{
Internal function. Not exported.
}
\author{
Sergei Tarasov
}
