% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhpp_functions.R
\name{KDE_unnorm_trunc_Markov}
\alias{KDE_unnorm_trunc_Markov}
\title{KDE for unnormalized Markov kernel vectorized.}
\usage{
KDE_unnorm_trunc_Markov(x, h, dat)
}
\description{
Internal function. Not exported.
}
\author{
Sergei Tarasov
}
