% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhpp_functions.R
\name{KDE_unnormalized_scalar_Markov_kernel}
\alias{KDE_unnormalized_scalar_Markov_kernel}
\title{KDE for unnormalized Markov kernel.}
\usage{
KDE_unnormalized_scalar_Markov_kernel(x, h, dat)
}
\arguments{
\item{x}{numeric. A vector with global changing times.}

\item{h}{numeric. The bandwidth value.

Internal function. Not exported.}

\item{dat}{numeric. A vector with path data values.}
}
\description{
KDE for unnormalized Markov kernel.
}
\author{
Sergei Tarasov
}
