% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhpp_functions.R
\name{make_data_NHPP_KDE_Markov_kernel}
\alias{make_data_NHPP_KDE_Markov_kernel}
\title{Get NHPP data for all edges (Markov KDE)}
\usage{
make_data_NHPP_KDE_Markov_kernel(Tb.trees)
}
\arguments{
\item{Tb.trees}{data.frame. A tibble obtained with the 'path_hamming_over_trees_KDE' function.}
}
\value{
A list with changing times between states for all edges of a given sample of trees.
}
\description{
Gets data on changing times between states for all edges of a given sample of trees for the Markov kernel density estimator (KDE).
}
\examples{
path.data <- make_data_NHPP_KDE_Markov_kernel(Tb.trees)

}
\author{
Sergei Tarasov
}
