% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_functions.R
\name{make_pic}
\alias{make_pic}
\title{Assign colors to picture ID layers}
\usage{
make_pic(picture, layers, cols.maps)
}
\arguments{
\item{picture}{Picture object.}

\item{layers}{numeric. A named vector where values indicate the layer IDs in the Picture object and names indicate the anatomy ontology term labels.}

\item{cols.maps}{character. A named vector where elements correspond to color IDs and names indicate the anatomy ontology term labels.}
}
\value{
An object of class 'Picture' with the assigned colors to different anatomical regions.
}
\description{
Assigns colors to picture ID layers (@paths) of an object of class 'Picture'.
The object should be a PS or ESP vector illustration imported using the grImport package.
Colors are taken from cols.maps argument were the palette indicates the scale of the desired statistics for the evolutionary rates.
}
\author{
Sergei Tarasov
}
