% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramo_functions.R
\name{merge_tree_cat_list}
\alias{merge_tree_cat_list}
\title{Merge state bins over a tree list}
\usage{
merge_tree_cat_list(tree.list)
}
\arguments{
\item{tree.list}{multiSimmap object.}
}
\value{
A list of trees with merged identical bins.
}
\description{
A wrapper function to merge identical state bins over a tree list.
}
\examples{
tree_list <- data(HYM)$stm[[1]]
tree_list <- discr_Simmap(tree_list)
stm_merg_list <- merge_tree_cat_list(tree_list)
stm_merg_list

}
\author{
Diego S. Porto
}
