% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphospace_functions.R
\name{MultiScale.simmap}
\alias{MultiScale.simmap}
\title{Multidimensional scaling of character states from one stochastic character map}
\usage{
MultiScale.simmap(tree.merge, add.noise = NULL)
}
\arguments{
\item{tree.merge}{simmap object. A stochastic character map after being processed by the discr_Simmap and merge_tree_cat functions.}

\item{add.noise}{numeric. A vector of length 2 or NULL. Indicates if noise should be added or not. Useful if there are many identical states occupying the same points in the 2D coordinates of the morphospace plot. 
The noise is calculated as var(V)*add.noise.}
}
\value{
A list of tibbles -- Points, Lines, and Edge.map -- correponding to tree branch information to plot.
}
\description{
Performs multidimensional scaling (MDS) based on hamming distances among character state vectors from one stochastic character map.
}
\examples{
data("hym_stm_mds")
# Get a sample of 5 amalgamated stochastic maps.
tree_list <- hym_stm_mds
tree_list <- merge_tree_cat_list(tree_list)
# Multidimensional scaling for an arbitrary tree.
\dontrun{
  
  MD <- suppressWarnings(MultiScale.simmap(tree_list[[1]], add.noise = c(0.3,0.3)))
  
}

}
\author{
Sergei Tarasov
}
