% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramo_functions.R
\name{paramo}
\alias{paramo}
\title{PARAMO}
\usage{
paramo(rac_query, tree.list, ntrees)
}
\arguments{
\item{rac_query}{character list. Named list obtained from the RAC_query function.}

\item{tree.list}{multiSimmap or multiPhylo object. Named list with stochastic character maps.}

\item{ntrees}{integer. Number of trees to stack.}
}
\value{
A list of stacked stochastic character maps.
}
\description{
Wrapper function to perform the final paramo stacking of maps for a set of anatomy ontology terms.
}
\examples{
tree_list <- data(HYM)$stm
tree_list <- lapply(tree_list, discr_Simmap)
tree_list <- lapply(tree_list, merge_tree_cat_list)
paramo.list(query, tree_list, 10)

}
\author{
Diego S. Porto
}
