% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhpp_functions.R
\name{posterior_lambda_KDE_Distr}
\alias{posterior_lambda_KDE_Distr}
\title{Get distributions of analytical posterior}
\usage{
posterior_lambda_KDE_Distr(tree.list, n.sim = 10, BR.name)
}
\arguments{
\item{tree.list}{multiSimmap object.}

\item{n.sim}{integer. Number of simulations.}

\item{BR.name}{character. A label name for the anatomical region.}
}
\value{
A tibble with the simulated distribution.
}
\description{
Simulates a distribution of number of state changes across all branches of all trees
}
\examples{
tree.list <- Trees.focal.BRs[[1]]
posterior_lambda_KDE_Distr(tree.list, n.sim = 10, "br")

}
\author{
Sergei Tarasov
}
