% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramo_functions.R
\name{stack2}
\alias{stack2}
\title{Stack two discrete stochastic character map lists; x and y are the list of state names (i.e. maps).}
\usage{
stack2(x, y)
}
\description{
Internal function. Not exported.
}
\author{
Sergei Tarasov
}
